A garbage sort of parser
It's interesting that pyparsing.QuotedString won't work. Basically, it will return the string without the escapes in it as content, so you don't know if 'a' in the output was 'a' or '!a' to begin with. And we want to count the former but not the latter. I wonder if they did that on purpose...